Move to using an extension file for specifying things like db
