+++ src/extensions.py
@@ -1,0 +1,25 @@
+"""
+Extensions module.
+
+Each extension is initialized in the app factory located in __init__.py
+
+We define the following extensions:
+
+- db: the SQLAlchemy ORM object (which is also used by Alembic for migrations)
+"""
+
+###################################################################################################
+#  Imports
+###################################################################################################
+
+from flask_sqlalchemy import SQLAlchemy
+
+###################################################################################################
+# Globals
+###################################################################################################
+
+db = SQLAlchemy()
+
+###################################################################################################
+#  End of file
+###################################################################################################
